continue transition to NHWC

+++ Demos/AlexNetDemo.cpp
@@ -28,7 +28,7 @@ int main() {
     assert(boost::filesystem::exists("/media/andrew/PCIE_SSD/ImageNet2012_2_of_2.shard"));
     shardPaths.insert("/media/andrew/PCIE_SSD/ImageNet2012_1_of_2.shard");
     shardPaths.insert("/media/andrew/PCIE_SSD/ImageNet2012_2_of_2.shard");
-    ThorImplementation::TensorDescriptor exampleDescriptor(ThorImplementation::TensorDescriptor::DataType::UINT8, {3, 224, 224});
+    ThorImplementation::TensorDescriptor exampleDescriptor(ThorImplementation::TensorDescriptor::DataType::FP16, {3, 224, 224});
 
     std::shared_ptr<LocalBatchLoader> batchLoader = make_shared<LocalBatchLoader>(shardPaths, exampleDescriptor, 512);
     batchLoader->setDatasetName("ImageNet 2012");
